requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.3
fpdf
hexdump==3.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
phonenumbers==9.0.26
pillow==12.1.1
psutil==7.2.2
pyaes==1.6.1
pyasn1==0.6.3
pyngrok==7.5.1
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.9.6
pytsk3==20250801
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scapy==2.7.0
selenium
six==1.17.0
soupsieve==2.8.3
Telethon==1.42.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
<<<<<<< HEAD
=======
opencage
>>>>>>> 9866c5a (Initial Private-Ready Release v1.2.2)
